// src/generated/notification.ts
export interface Notification {
  id: string;
  postId: string;
  creatorId: string;
  creatorAvtUrl: string;
  receiverId: string;
  notifType: NotificationType;
  message: string;
  isRead: boolean;
  createdAt: string;
  deletedAt: string;
}

export interface GetNotificationsRequest {
  userId: string;
}

export interface GetNotificationsResponse {
  notifications: Notification[];
}

export interface CreateNotificationRequest {
  postId: string;
  creatorId: string;
  creatorAvtUrl: string;
  receiverId: string;
  notifType: NotificationType;
  message: string;
}

export interface CreateNotificationResponse {
  notifications: Notification | undefined;
}

export interface DeleteNotificationsRequest {
  id: string;
}

export interface DeleteNotificationsResponse {
  notifications: Notification | undefined;
}

export interface MarkAsReadNotificationRequest {
  id: string;
}

export interface MarkAsReadotificationResponse {
  notification: Notification | undefined;
}

export interface MarkAsReadNotificationsRequest {
  ids: string[];
}

export interface MarkAsReadotificationsResponse {
  notifications: Notification[];
}

export const NOTIFICATION_PACKAGE_NAME = "notification";

function createBaseNotification(): Notification {
  return {
    id: "",
    postId: "",
    creatorId: "",
    creatorAvtUrl: "",
    receiverId: "",
    notifType: 0,
    message: "",
    isRead: false,
    createdAt: "",
    deletedAt: "",
  };
}

export const Notification: MessageFns<Notification> = {
  encode(message: Notification, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.postId !== "") {
      writer.uint32(18).string(message.postId);
    }
    if (message.creatorId !== "") {
      writer.uint32(26).string(message.creatorId);
    }
    if (message.creatorAvtUrl !== "") {
      writer.uint32(34).string(message.creatorAvtUrl);
    }
    if (message.receiverId !== "") {
      writer.uint32(42).string(message.receiverId);
    }
    if (message.notifType !== 0) {
      writer.uint32(48).int32(message.notifType);
    }
    if (message.message !== "") {
      writer.uint32(58).string(message.message);
    }
    if (message.isRead !== false) {
      writer.uint32(64).bool(message.isRead);
    }
    if (message.createdAt !== "") {
      writer.uint32(74).string(message.createdAt);
    }
    if (message.deletedAt !== "") {
      writer.uint32(82).string(message.deletedAt);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Notification {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseNotification();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.postId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.creatorId = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.creatorAvtUrl = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.receiverId = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.notifType = reader.int32() as any;
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 8: {
          if (tag !== 64) {
            break;
          }

          message.isRead = reader.bool();
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.createdAt = reader.string();
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.deletedAt = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const GetNotificationsResponse: MessageFns<GetNotificationsResponse> = {
  encode(message: GetNotificationsResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.notifications) {
      Notification.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetNotificationsResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetNotificationsResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.notifications.push(Notification.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseCreateNotificationRequest(): CreateNotificationRequest {
  return { postId: "", creatorId: "", creatorAvtUrl: "", receiverId: "", notifType: 0, message: "" };
}

export const CreateNotificationRequest: MessageFns<CreateNotificationRequest> = {
  encode(message: CreateNotificationRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.postId !== "") {
      writer.uint32(18).string(message.postId);
    }
    if (message.creatorId !== "") {
      writer.uint32(26).string(message.creatorId);
    }
    if (message.creatorAvtUrl !== "") {
      writer.uint32(34).string(message.creatorAvtUrl);
    }
    if (message.receiverId !== "") {
      writer.uint32(42).string(message.receiverId);
    }
    if (message.notifType !== 0) {
      writer.uint32(48).int32(message.notifType);
    }
    if (message.message !== "") {
      writer.uint32(58).string(message.message);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateNotificationRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateNotificationRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.postId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.creatorId = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.creatorAvtUrl = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.receiverId = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.notifType = reader.int32() as any;
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.message = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const DeleteNotificationsRequest: MessageFns<DeleteNotificationsRequest> = {
  encode(message: DeleteNotificationsRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DeleteNotificationsRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDeleteNotificationsRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseDeleteNotificationsResponse(): DeleteNotificationsResponse {
  return { notifications: undefined };
}

export const DeleteNotificationsResponse: MessageFns<DeleteNotificationsResponse> = {
  encode(message: DeleteNotificationsResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.notifications !== undefined) {
      Notification.encode(message.notifications, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DeleteNotificationsResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDeleteNotificationsResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.notifications = Notification.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};

function createBaseMarkAsReadNotificationRequest(): MarkAsReadNotificationRequest {
  return { id: "" };
}

export const MarkAsReadNotificationRequest: MessageFns<MarkAsReadNotificationRequest> = {
  encode(message: MarkAsReadNotificationRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MarkAsReadNotificationRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMarkAsReadNotificationRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export const MarkAsReadNotificationsRequest: MessageFns<MarkAsReadNotificationsRequest> = {
  encode(message: MarkAsReadNotificationsRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.ids) {
      writer.uint32(10).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MarkAsReadNotificationsRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMarkAsReadNotificationsRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.ids.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
}